cl/stopwatch: make the timing info more precise
... by using floats at certain places to avoid integral truncation.

+++ cl/stopwatch.hh
@@ -28,7 +28,7 @@ class StopWatch {
         ~StopWatch();
 
         void reset();
-        long /* msec */ elapsed() const;
+        float /* sec */ elapsed() const;
 
     private:
         struct Private;
